+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   point). This adapter is needed when serving ES5 to browsers that support the
   native Custom Elements API
   ([#164](https://github.com/Polymer/polyserve/issues/164)).
+* A server can now be started with a custom entry point; previously index.html
+  was always assumed ([#161](https://github.com/Polymer/polyserve/issues/161)).
 
 ## [0.17.0](https://github.com/PolymerLabs/polyserve/tree/v0.17.0) (2017-04-13)